# arch-register-packages/server/openapi.yaml
openapi: 3.1.0
info:
  title: Arch Register API
  version: 1.0.0
  description: REST API for the Architecture Register.
servers:
  - url: http://localhost:3010
    description: Local development server
...
paths:
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      operationId: listWorkspaces
      responses:
        '200':
          description: Workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Workspaces]
      summary: Create workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '200':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with username and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/auth/oidc/authorize:
    get:
      tags: [Authentication]
      summary: Initiate OIDC authentication flow
      operationId: oidcAuthorize
      responses:
        '200':
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCAuthorizeResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/auth/oidc/callback:
    get:
      tags: [Authentication]
      summary: OIDC callback endpoint
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to frontend with auth cookies set
        '400':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New authentication tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user information
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/auth/users:
    get:
      tags: [Authentication]
      summary: List all users (admin only)
      operationId: listUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceUserInfo'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/users/{id}:
    patch:
      tags: [Authentication]
      summary: Update current user's account settings
      operationId: updateUserAccount
      description: Allows users to update their own account settings. Users can only update their own account.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID (must match authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAccountRequest'
      responses:
        '200':
          description: User account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceUserInfo'
        '403':
          description: Forbidden - can only update own account
          $ref: '#/components/responses/Error'
        '404':
          description: User not found
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'


      tags: [Authentication]
      summary: Get user's global roles
      operationId: getUserGlobalRoles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User's global roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GlobalRoleAssignment'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Authentication]
      summary: Update user's global roles
      operationId: updateUserGlobalRoles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGlobalRolesRequest'
      responses:
        '200':
          description: Updated global roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GlobalRoleAssignment'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        statusText:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/ArbitraryJson'
      additionalProperties: true
...
    ArbitraryJson:
      oneOf:
        - type: object
          additionalProperties:
            $ref: '#/components/schemas/ArbitraryJson'
        - type: array
          items:
            $ref: '#/components/schemas/ArbitraryJson'
        - type: string
        - type: number
        - type: integer
        - type: boolean
        - type: 'null'
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url_slug:
          type: string
        short_code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, url_slug, short_code, description, created_at, updated_at]
    CreateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
      additionalProperties: false
...
    WorkspaceUserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        display_name:
          type: string
        auth_provider:
          type: string
        is_active:
          type: boolean
        color:
          type: [string, 'null']
      required: [id, email, display_name, auth_provider, is_active]
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
      required: [access_token, refresh_token, token_type, expires_in]
    OIDCAuthorizeResponse:
      type: object
      properties:
        authorization_url:
          type: string
      required: [authorization_url]
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
    CurrentUserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        display_name:
          type: string
        auth_provider:
          type: string
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: [string, 'null']
          format: date-time
        global_roles:
          type: array
          items:
            type: string
        global_permissions:
          type: array
          items:
            type: string
        team_assignments_by_workspace:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                team_id:
                  type: string
                role:
                  type: string
              required: [team_id, role]
        workspace_roles:
          type: object
          additionalProperties:
            type: string
        teams_by_workspace:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                type:
                  type: string
                  const: team
              required: [id, name, type]
      required: [id, email, display_name, auth_provider, created_at, last_login_at, global_roles, global_permissions, team_assignments_by_workspace, workspace_roles, teams_by_workspace]
    GlobalRoleAssignment:
      type: object
      properties:
        user_id:
          type: string
        role:
          type: string
          enum: [global_admin, workspace_admin]
        created_at:
          type: string
          format: date-time
      required: [user_id, role, created_at]
    UpdateGlobalRolesRequest:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
            enum: [global_admin, workspace_admin]
      required: [roles]
...
    UpdateUserAccountRequest:
      type: object
      properties:
        color:
          type: [string, 'null']
          description: User's preferred color for avatar/display
        display_name:
          type: string
          description: User's display name
      additionalProperties: false