# arch-register-packages/server/src/types.ts
export type TextField = {
  id: string;
  name: string;
  type: 'text' | 'longtext';
};

export type BooleanField = {
  id: string;
  name: string;
  type: 'boolean';
};

export type SelectField = {
  id: string;
  name: string;
  type: 'select';
  options: Array<{ value: string; label: string }>;
};

// Points to another entity of a given schema type; multiple values stored as comma-separated UUIDs.
// maxCount: -1 means unbounded.
export type ReferenceField = {
  id: string;
  name: string;
  type: 'reference';
  schemaId: string;
  minCount: number;
  maxCount: number;
};

// Like ReferenceField but expresses a parent/child containment relationship.
// The child entity stores the parent's UUID in its data.
export type ContainmentField = {
  id: string;
  name: string;
  type: 'containment';
  schemaId: string;
  minCount: number;
  maxCount: number;
};

export type SchemaField = TextField | BooleanField | SelectField | ReferenceField | ContainmentField;

export type Workspace = {
  id: string;
  name: string;
  url_slug: string;
  short_code: string;
  description: string;
  created_at: Date;
  updated_at: Date;
};

export type EntitySchema = {
  id: string;
  workspace: string;
  name: string;
  fields: SchemaField[];
  color: string | null;
  icon: string | null;
  default_owner: string | null;
  created_at: Date;
  updated_at: Date;
};

export type LifecycleStatus = string;

export type WorkspaceLifecycleState = {
  id: string;
  workspace: string;
  label: string;
  color: string;
  sort_order: number;
  created_at: Date;
};

export type WorkspaceOwner = {
  id: string;
  workspace: string;
  sort_order: number;
  color: string | null;
  description: string;
  created_at: Date;
};

export type EntityLink = {
  url: string;
  title: string;
  type?: string;
};

export type VisibilityMode = 'public' | 'restricted';

export type Entity = {
  id: string;
  workspace: string;
  slug: string;
  namespace: string;
  name: string;
  description: string;
  owner: string | null;
  lifecycle: LifecycleStatus | null;
  tags: string[];
  links: EntityLink[];
  schema_id: string;
  data: Record<string, unknown>;
  visibility_mode: VisibilityMode | null;
  created_at: Date;
  updated_at: Date;
};

// Wire format returned by the API: first-class metadata prefixed with _ to avoid clashing with data fields.
export type EntityApiResponse = {
  _uid: string;
  _workspace: string;
  _schemaId: string;
  _name: string;
  _slug: string;
  _namespace: string;
  _description: string;
  _owner: string | null;
  _lifecycle: LifecycleStatus | null;
  _tags: string[];
  _links: EntityLink[];
  _visibilityMode: VisibilityMode | null;
  [field: string]: unknown;
};

export type Project = {
  id: string;
  workspace: string;
  name: string;
  description: string;
  owner: string | null;
  status: 'pinned' | 'active' | 'archived';
  color: string | null;
  created_at: Date;
  updated_at: Date;
};

export type ProjectFile = {
  id: string;
  workspace: string;
  project_id: string;
  path: string;
  name: string;
  size_bytes: number;
  is_template: boolean;
  is_workspace_template: boolean;
  created_at: Date;
  updated_at: Date;
};

export const encodeRefs = (refs: string[]): string => refs.join(',');
export const decodeRefs = (raw: unknown): string[] => {
  if (raw == null || raw === '') return [];
  return String(raw)
    .split(',')
    .map(s => s.trim())
    .filter(Boolean);
};

export type AuditOperation = 'create' | 'update' | 'delete';

export type AuditEntityType = 'workspace' | 'entity_schema' | 'entity' | 'project' | 'project_file';

export type AuditLogEntry = {
  id: string;
  workspace: string;
  timestamp: Date;
  user_id: string;
  operation: AuditOperation;
  entity_type: AuditEntityType;
  entity_id: string;
  entity_name: string;
  entity_slug: string | null;
  schema_id: string | null;
  changes: {
    old?: Record<string, unknown>;
    new?: Record<string, unknown>;
  };
  metadata: Record<string, unknown>;
};

export type AuditLogApiResponse = {
  id: string;
  workspace: string;
  timestamp: string;
  user_id: string;
  operation: AuditOperation;
  entity_type: AuditEntityType;
  entity_id: string;
  entity_name: string;
  entity_slug: string | null;
  schema_id: string | null;
  changes: {
    old?: Record<string, unknown>;
    new?: Record<string, unknown>;
  };
  metadata: Record<string, unknown>;
};

export type AuthProvider = 'local' | 'oidc';

export type User = {
  id: string;
  email: string | null;
  display_name: string;
  auth_provider: AuthProvider;
  password_hash: string | null;
  oidc_issuer: string | null;
  oidc_subject: string | null;
  is_active: boolean;
  color: string | null;
  created_at: Date;
  updated_at: Date;
  last_login_at: Date | null;
};

export type GlobalRole = 'global_admin' | 'workspace_admin';

export type GlobalPermission =
  | 'admin_platform'
  | 'create_workspaces'
  | 'manage_workspace_roles';

export type WorkspaceRole = 'owner' | 'admin' | 'editor' | 'reviewer' | 'viewer';

export type TeamRole = 'team_admin' | 'team_editor' | 'team_reviewer';

export type WorkspaceMember = {
  workspace: string;
  user_id: string;
  role: WorkspaceRole;
  created_at: Date;
};

export type TeamMembership = {
  workspace: string;
  team_id: string;
  user_id: string;
  role: TeamRole;
  created_at: Date;
};

export type GlobalRoleAssignment = {
  user_id: string;
  role: GlobalRole;
  created_at: Date;
};

export type EntityRole = 'viewer' | 'editor' | 'contributor' | 'entity_admin';
export type EntityGrantScope = 'self' | 'subtree';

export type EntityCapabilities = {
  canView: boolean;
  canEdit: boolean;
  canDelete: boolean;
  canAdmin: boolean;
  canCreateChild: boolean;
};

export type ProjectCapabilities = {
  canEdit: boolean;
  canDelete: boolean;
  canManageFiles: boolean;
};

export type EntityGrant = {
  id: string;
  workspace: string;
  entity_id: string;
  principal_type: 'user' | 'team';
  principal_id: string;
  role: EntityRole;
  applies_to: EntityGrantScope;
  created_at: Date;
};

export type JWTPayload = {
  sub: string;
  email?: string;
  name: string;
  provider: AuthProvider;
  type: 'access' | 'refresh';
  iat: number;
  exp: number;
};

export type AuthContext = {
  user: User;
  token: string;
};